% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{ci.fisher}
\alias{ci.fisher}
\title{Fisher confidence interval}
\usage{
ci.fisher(alpha, cor, se)
}
\arguments{
\item{alpha}{alpha value for 1-alpha confidence}

\item{cor}{estimated correlation or association coefficient}

\item{se}{standard error of estimate}
}
\value{
Returns a 1-row matrix containing the lower and upper confidence limits.
}
\description{
Computes a Fisher confidence interval for any type of correlation or any
measure of association that has a -1 to 1 range.
}
\examples{
ci.fisher(.05, .641, .052)

# Should return:
#             LL        UL
# [1,] 0.5276396 0.7319293


}
