% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{ci.kendalltau}
\alias{ci.kendalltau}
\title{Confidence interval for a Kendall tau-a correlation}
\usage{
ci.kendalltau(alpha, cor, n)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{cor}{estimate of Kendall's tau-a correlation}

\item{n}{sample size}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item Estimate - estimated Kendall tau-a correlation (from input)
\item SE - standard error
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
Computes a Fisher confidence interval for a population Kendal tau-a correlation.
This function is useful for Kendall tau-a correlations less than about .8. (Note:
under bivariate normality, a Kendall tau-a correlation of .8 corresponds to a
Pearson correlation of about .95.) An estimate of the Kendall tau-a correlation
can be obtained from the (see \link[stats]{cor.test} function with method = "kendall".
}
\examples{
ci.kendalltau(.05, .734, 40)

# Should return:
# Estimate      SE     LL     UL
#    0.734 0.05082 0.6178 0.8188


}
\references{
\insertRef{Bonett2002}{statpsych}
}
