% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{ci.reliability}
\alias{ci.reliability}
\title{Confidence interval for a reliability coefficient}
\usage{
ci.reliability(alpha, rel, se, n)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{rel}{estimated reliability}

\item{se}{standard error of reliability}

\item{n}{sample size}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item Estimate - estimated reliability (from input)
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
Computes a confidence interval for a population reliability coefficient
such as Cronbach's alpha or McDonald's omega using an estimate of the
reliability and its standard error. The standard error can be a robust
standard error or bootstrap standard error obtained from an SEM program.
Use \link[statpsych]{ci.cronbach} for Cronbach's alpha if parallel
measurements can be assumed.
}
\examples{
ci.reliability(.05, .88, .0147, 100)

# Should return:
# Estimate         LL        UL
#     0.88  0.8489612 0.9065575
 

}
