% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{etasqr.adj}
\alias{etasqr.adj}
\title{Bias adjustment for an eta-squared estimate}
\usage{
etasqr.adj(etasqr, dfeffect, dferror)
}
\arguments{
\item{etasqr}{unadjusted eta-square estimate}

\item{dfeffect}{degrees of freedom for the effect}

\item{dferror}{error degrees of freedom}
}
\value{
Returns a bias adjusted eta-squared estimate
}
\description{
Computes an approximate bias adjustment for eta-squared. This adjustment
can be applied to eta-squared, partial-eta squared, and generalized
eta-squared estimates.

For more details, see Section 3.7 of Bonett (2021, Volume 1)
}
\examples{
etasqr.adj(.315, 2, 42)

# Should return:
# adj Eta-squared
#          0.2824
 

}
\references{
\insertRef{Bonett2021}{statpsych}
}
