% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{iqv}
\alias{iqv}
\title{Indices of qualitative variation}
\usage{
iqv(f)
}
\arguments{
\item{f}{vector of multinomial frequency counts}
}
\value{
Returns estimates of the Shannon, Berger, and Simpson qualitative indices
}
\description{
Computes the Shannon, Berger, and Simpson indices of qualitative variation.
}
\examples{
f <- c(10, 46, 15, 3)
iqv(f)
# Should return:
#        Simpson    Berger   Shannon
# [1,] 0.7367908 0.5045045       0.7
 

}
