% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{logitfit}
\alias{logitfit}
\title{The confusion matrix is computed using the observed 0 or 1 response variable
scores, the predicted probabilities from the logistic model, and a specified
cutpoint.}
\usage{
logitfit(y, p, cut)
}
\arguments{
\item{y}{vector of observed 0 or 1 response scores}

\item{p}{vector of predicted probabilities from model}

\item{cut}{cutpoint (defines the predicted 0 or 1 scores)}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item C - percent correctly classified (aka accuracy)
\item TP - percent true positives (aka sensitivity, recall)
\item FP - percent false positives (1 - TN)
\item TN - percent true negatives (aka specificity)
\item FN - percent false negatives (1 - TP)
\item PPV - percent positive predicted values (aka precision)
\item NPV - percent negative predicted values
\item F1 - F1 score (2 x PPV x TP)/(PPV + TP)
}
}
\description{
The confusion matrix is computed using the observed 0 or 1 response variable
scores, the predicted probabilities from the logistic model, and a specified
cutpoint.
}
\examples{
y <- c(0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)
x1 <- c(1,1,3,2,6,8,4,5,6,2,4,3,1,5,3,9,8,9,8,6,6,7,5,3,8,6,5,7,8,9,7,8)
x2 <- c(0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,0,0,0)
out <- glm(y ~ x1 + x2, family = binomial(link = "logit"))
p <- predict(out, type = "response")
logitfit(y, p, .3)

# Should return:
#      C    TP    FP    TN   FN PPV     NPV      F1 
#  81.25 93.75 31.25 68.75 6.25  75 91.6667 83.3333 

logitfit(y, p, .4)
# Should return:
#     C    TP    FP    TN   FN     PPV     NPV      F1
#  87.5 93.75 18.75 81.25 6.25 83.3333 92.8571 88.2353


}
