% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{power.cor}
\alias{power.cor}
\title{Approximates the power of a correlation test for a planned sample size}
\usage{
power.cor(alpha, n, cor, h, s)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{n}{planned sample size}

\item{cor}{planning value of correlation}

\item{h}{null hypothesis value of correlation}

\item{s}{number of control variables}
}
\value{
Returns the approximate power of the test
}
\description{
Computes the approximate power of a test for a population Pearson or partial
correlation test for a planned sample size. Set s = 0 for a Pearson correlation.
}
\examples{
power.cor(.05, 80, .3, 0, 0)

# Should return:
#  Power
# 0.7752


}
