% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{power.cor1}
\alias{power.cor1}
\title{Approximates the power of a correlation test for a planned sample size}
\usage{
power.cor1(alpha, n, cor, h, s)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{n}{planned sample size}

\item{cor}{planning value of correlation}

\item{h}{hypothesized value of correlation}

\item{s}{number of control variables}
}
\value{
Returns the approximate power of the test
}
\description{
Computes the approximate power of a Pearson or partial correlation test for
a planned sample size. Set s = 0 for a Pearson correlation.
}
\examples{
power.cor1(.05, 80, .3, 0, 0)

# Should return:
#          Power
# [1,] 0.7751932


}
