% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{power.mann}
\alias{power.mann}
\title{Approximates the power of a Mann-Whitney test}
\usage{
power.mann(alpha, n1, n2, es)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{n1}{planned sample size for group 1}

\item{n2}{planned sample size for group 2}

\item{es}{planning value of effect size (between .5 and 1)}
}
\value{
Returns the approximate power of the test
}
\description{
Computes the approximate power of a Mann-Whitney U test for a planned
sample size. In a 2-group experiment, the effect size is the proportion
of members in the population with scores that would be higher under
treatment 1 than treatment 2. In a 2-group nonexperiment where participants
are sampled from two subpopulations of sizes N1 and N2, the effect size is
the proportion of all N1 x N2 pairs in which a member from subpopulation 1
has a larger score than a member from subpopulation 2.
}
\examples{
power.mann(.05, 50, 50, .65)

# Should return:
#   Power
#  0.7383


}
\references{
\insertRef{Noether1987}{statpsych}
}
