% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{random.sample}
\alias{random.sample}
\title{Generate a random sample}
\usage{
random.sample(popsize, samsize)
}
\arguments{
\item{popsize}{study population size}

\item{samsize}{sample size}
}
\value{
Returns a vector of randomly generated participant IDs
}
\description{
Generates a random sample of participant IDs without replacement.
}
\examples{
random.sample(3000, 25)

# Should return:
#  [1]   37   94  134  186  212  408  485  697  722  781  998 1055 
# [13] 1182 1224 1273 1335 1452 1552 1783 1817 2149 2188 2437 2850 2936
 

}
