% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{randomize}
\alias{randomize}
\title{Randomize a sample into groups}
\usage{
randomize(n)
}
\arguments{
\item{n}{vector of sample sizes per group}
}
\value{
Returns a vector of randomly generated group assignments.
}
\description{
Randomly assigns a sample of participants into k groups.
}
\examples{
n <- c(10, 10, 5)
randomize(n)

# Should return:
# [1] 2 3 2 1 1 2 3 3 2 1 2 1 3 1 1 2 3 1 1 2 2 1 1 2 2
 

}
