% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.ci.cv}
\alias{size.ci.cv}
\title{Sample size for a coefficient of variation}
\usage{
size.ci.cv(alpha, CV, w)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{CV}{planning value of coefficient of variation}

\item{w}{desired confidence interval width}
}
\value{
Returns the required sample size
}
\description{
Computes an approximate sample size required to estimate a population
coefficient of variation (CV) with desired confidence interval precision.
Set the CV planning value to the largest value within a plausible range
for a conservatively large sample size.
}
\examples{
size.ci.cv(.05, .25, .10)

# Should return:
# Sample size
#          60
 

}
