% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{size.ci.kendalltau}
\alias{size.ci.kendalltau}
\title{Sample size for a Kendall-a correlation confidence interval}
\usage{
size.ci.kendalltau(alpha, cor, w)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{cor}{planning value of correlation}

\item{w}{desired confidence interval width}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to estimate a population Kendall tau-a
correlation with desired confidence interval precision. Set the correlation
planning value to the smallest absolute value within a plausible range for a
conservatively large sample size.
}
\examples{
size.ci.kendalltau(.05, .3, .2)

# Should return:
# Sample size
#         143
 

}
\references{
\insertRef{Bonett2000}{statpsych}
}
