% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.ci.mean1}
\alias{size.ci.mean1}
\title{Sample size for a single mean confidence interval}
\usage{
size.ci.mean1(alpha, var, w)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{var}{planning value of response variable variance}

\item{w}{desired confidence interval width}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to estimate a single population mean with
desired confidence interval precision. Set the variance planning value to
the largest value within a plausible range for a conservatively large
sample size.
}
\examples{
size.ci.mean1(.05, 264.4, 10)

# Should return:
# Sample size
#          43
 

}
