% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{size.ci.oddsratio}
\alias{size.ci.oddsratio}
\title{Sample size for an odds ratio confidence interval}
\usage{
size.ci.oddsratio(alpha, p1, p2, or, r)
}
\arguments{
\item{alpha}{alpha level for 1 - alpha confidence}

\item{p1}{planning value for row 1 marginal proportion}

\item{p2}{planning value for column 1 marginal proportion}

\item{or}{planning value of odds ratio}

\item{r}{desired upper to lower confidence interval endpoint ratio}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to estimate an odds ratio with desired
confidence interval precision.
}
\examples{
size.ci.oddsratio(.05, .3, .2, 5.5, 3.0)

# Should return:
#  Sample size
#          356


}
