% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{size.ci.prop}
\alias{size.ci.prop}
\title{Sample size for a proportion confidence interval}
\usage{
size.ci.prop(alpha, p, w)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{p}{planning value of proportion}

\item{w}{desired confidence interval width}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to estimate a population proportion
with desired confidence interval precision. Set the proportion planning
value to .5 for a conservatively large sample size.

For more details, see Section 1.14 of Bonett (2021, Volume 3)
}
\examples{
size.ci.prop(.05, .5, .1)

# Should return:
# Sample size
#         385


}
\references{
\insertRef{Bonett2021}{statpsych}
}
