% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{size.ci.prop1}
\alias{size.ci.prop1}
\title{Sample size for a single proportion confidence interval}
\usage{
size.ci.prop1(alpha, p, w)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{p}{planning value of proportion}

\item{w}{desired confidence interval width}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to estimate a single proportion with
desired confidence interval precision. Set the proportion planning value
to .5 for a conservatively large sample size.
}
\examples{
size.ci.prop1(.05, .4, .2)

# Should return:
# Sample size
#          93


}
