% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.ci.sd}
\alias{size.ci.sd}
\title{Sample size for a standard deviation confidence interval}
\usage{
size.ci.sd(alpha, r)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{r}{desired upper to lower confidence interval endpoint ratio}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to estimate a population standard
deviation with desired confidence interval precision. This function
assumes that the traditional confidence interval for a population
standard deviation will be used. The traditional confidence interval
assumes that the response variable has an approximate normal
distribution and can be highly inaccurate when this assumption is not
satisfied.
}
\examples{
size.ci.sd(.05, 1.5)

# Should return:
# Sample size
#          49
 

}
