% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.test.cronbach}
\alias{size.test.cronbach}
\title{Sample size to test a Cronbach reliability}
\usage{
size.test.cronbach(alpha, pow, rel, r, h)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{rel}{reliability planning value}

\item{r}{number of measurements}

\item{h}{null hypothesis value of reliability}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to test a Cronbach reliability with
desired power.
}
\examples{
size.test.cronbach(.05, .85, .80, 5, .7)

# Should return:
# Sample size
#         139
 

}
\references{
\insertRef{Bonett2015}{statpsych}
}
