% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{size.test.cronbach2}
\alias{size.test.cronbach2}
\title{Sample size to test equality of Cronbach reliability coefficients in a
2-group design}
\usage{
size.test.cronbach2(alpha, pow, rel1, rel2, r)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{rel1}{group 1 reliability planning value}

\item{rel2}{group 2 reliability planning value}

\item{r}{number of measurements (items, raters)}
}
\value{
Returns the required sample size for each group
}
\description{
Computes the sample size required to test a difference in population
Cronbach reliability coefficients with desired power in a 2-group design.
}
\examples{
size.test.cronbach2(.05, .80, .85, .70, 8)

# Should return:
# Sample size per group
#                    77
 

}
\references{
\insertRef{Bonett2015}{statpsych}
}
