% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.test.icc}
\alias{size.test.icc}
\title{Sample size to test an intraclass correlation}
\usage{
size.test.icc(alpha, pow, icc, r, h)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{icc}{icc planning value}

\item{r}{number of measurements (items, raters, forms)}

\item{h}{null hypothesis value of icc}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to test an intraclass correlation (icc)
with desired power in a one-way random effects ANOVA model.
}
\examples{
size.test.icc(.05, .90, .65, 4, .50)

# Should return:
# Sample size
#         104
 

}
\references{
\insertRef{Donner1987}{statpsych}
}
