% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.test.mean1}
\alias{size.test.mean1}
\title{Sample size for a test of a single mean}
\usage{
size.test.mean1(alpha, pow, var, es)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{var}{planning value of response variable variance}

\item{es}{planning value of mean minus null hypothesis value}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to test a single population mean with
desired power in a 1-group design. Set the variance planning value to the
largest value within a plausible range for a conservatively large sample
size.
}
\examples{
size.test.mean1(.05, .9, 80.5, 7)

# Should return:
# Sample size
#          20
 

}
