% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{size.test.prop1}
\alias{size.test.prop1}
\title{Sample size for a test of a single proportion}
\usage{
size.test.prop1(alpha, pow, p, es)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{p}{planning value of proportion}

\item{es}{planning value of proportion minus null hypothesis value}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to test a single population proportion
with desired power in a 1-group design. Set the proportion planning value
to .5 for a conservatively large sample size. The value of the effect
size (expected population proportion minus hypothesized value) need not
be based on the proportion planning value.
}
\examples{
size.test.prop1(.05, .9, .5, .2)

# Should return:
#      Sample size
# [1,]          66


}
