% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{size.test.sign1}
\alias{size.test.sign1}
\title{Sample size for a 1-sample Sign test}
\usage{
size.test.sign1(alpha, pow, p)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{p}{planning value of proportion}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required for a Sign test with desired power in a
1-sample design. A planning value of the 1-sample Sign test parameter value
is required. This parameter is the proportion of members in the population
with scores greater than the hypothesized value.
}
\examples{
size.test.sign1(.05, .90, .3)

# Should return:
# Sample size
#          56
 

}
