% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{slope.contrast}
\alias{slope.contrast}
\title{Contrast coefficients for the slope of a quantitative factor}
\usage{
slope.contrast(x)
}
\arguments{
\item{x}{vector of numeric factor levels}
}
\value{
Returns the vector of contrast coefficients
}
\description{
Computes the contrast coefficients that are needed to estimate the slope of
a line in a one-factor design with a quantitative factor.
}
\examples{
x <- c(25, 50, 75, 100)
slope.contrast(x)

# Should return:
#      Coefficient
# [1,]      -0.012
# [2,]      -0.004
# [3,]       0.004
# [4,]       0.012
 

}
