% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{spearmanbrown}
\alias{spearmanbrown}
\title{Computes the reliability of a scale with r2 measurements given the
reliability of a scale with r1 measurements}
\usage{
spearmanbrown(rel, r1, r2)
}
\arguments{
\item{rel}{reliability of the sum or average of r1 measurements}

\item{r1}{number of measurements in the original scale}

\item{r2}{number of measurements in the new scale}
}
\value{
Returns the reliability of the sum or average of r2 measurements
}
\description{
Computes the reliability of a scale that is the sum or average of r2
parallel measurements given the reliability of a scale that is the sum or
average of r1 parallel measurements. The "measurements" can be items,
forms, raters, or occasions.

For more details, see Section 4.19 of Bonett (2021, Volume 1)
}
\examples{
spearmanbrown(.6, 10, 20)

# Should return:
# Reliability of 20 measurements
#                            .75


}
\references{
\insertRef{Bonett2021}{statpsych}
}
