% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{test.kurtosis.geary}
\alias{test.kurtosis.geary}
\title{Computes estimate and test of excess Geary kurtosis}
\usage{
test.kurtosis.geary(y)
}
\arguments{
\item{y}{vector of quantitative scores}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item Kurtosis - estimate of transformed Geary kurtosis coefficient
\item Excess - estimate of excess kurtosis (kurtosis - 3)
\item z - z test statistic
\item p - two-sided p-value
}
}
\description{
Computes an estimate and test for kurtosis using a modfication of Geary's
measure of kurtosis. If the p-value is small (e.g., less than .05) and excess
kurtosis is positive, then the normality assumption can be rejected due to
leptokurtosis. If the p-value is small (e.g., less than .05) and excess
kurtosis is negative, then the normality assumption can be rejected due to
platykurtosis. The estimate and test of Geary's kurtosis used here is based
on a transformation of Geary's orginal measure of kurtosis proposed by
Bonett and Seier (2002). Geary's kurtosis tends to be more sensitive to
peakedness than Pearson's kurtosis, and Pearson's kurtosis tend to be more
sensitive to tail weight than Geary's kurtosis. In the same way that it is
informative to assess centrality and variability using more than one
measure, it is also informative to assess kurtosis using both Pearson
kurtosis and Geary kurtosis. See (see \link[statpsych]{test.kurtosis}) for
a test of Pearson kurtosis.
}
\examples{
y <- c(58, 58, 55, 52, 20, 65, 59, 49, 51, 81, 40, 62, 56, 49, 49, 50, 44, 53, 59)
test.kurtosis.geary(y)

# Should return:
# Kurtosis Excess     z     p
#    5.157  2.157 2.793 0.005


}
\references{
\insertRef{Bonett2002c}{statpsych}
}
