% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{test.kurtosis}
\alias{test.kurtosis}
\title{Computes p-value for test of excess kurtosis}
\usage{
test.kurtosis(y)
}
\arguments{
\item{y}{vector of quantitative scores}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item Kurtosis - estimate of kurtosis coefficient
\item Excess - estimate of excess kurtosis (kurtosis - 3)
\item p - Monte Carlo two-sided p-value
}
}
\description{
Computes a Monte Carlo p-value (250,000 replications) for the null
hypothesis that the sample data come from a normal distribution. If the
p-value is small (e.g., less than .05) and excess kurtosis is positive,
then the normality assumption can be rejected due to leptokurtosis. If the
p-value is small (e.g., less than .05) and excess kurtosis is negative,
then the normality assumption can be rejected due to platykurtosis.

For more details, see Section 1.23 of Bonett (2021, Volume 1)
}
\examples{
y <- c(58, 58, 55, 52, 20, 65, 59, 49, 51, 81, 40, 62, 56, 49, 49, 50, 44, 53, 59)
test.kurtosis(y)

# Should return:
# Kurtosis Excess      p
#   5.5081 2.5081 0.0164


}
\references{
\insertRef{Bonett2021}{statpsych}
}
