% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{test.prop1}
\alias{test.prop1}
\title{Hypothesis test for a single proportion}
\usage{
test.prop1(f, n, h)
}
\arguments{
\item{f}{number of participants who have the attribute}

\item{n}{sample size}

\item{h}{population proportion under null hypothesis}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item Estimate - ML estimate of proportion
\item z - z test statistic
\item p - p-value
}
}
\description{
Computes a continuity-corrected z test for a single proportion in a
1-group design.
}
\examples{
test.prop1(9, 20, .2)

# Should return:
# Estimate        z          p
#     0.45 2.515576 0.01188379


}
\references{
\insertRef{Snedecor1980}{statpsych}
}
