% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{ci.biphi}
\alias{ci.biphi}
\title{Confidence interval for a biserial-phi correlation}
\usage{
ci.biphi(alpha, f1, f2, n1, n2)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{f1}{number of participants in group 1 who have the attribute}

\item{f2}{number of participants in group 2 who have the attribute}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}
}
\value{
Returns a 1-row matrix. The columns are:
\itemize{
\item Estimate - estimate of biserial-phi correlation
\item SE - standard error
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
Computes a confidence interval for a biserial-phi correlation using a
transformation of a confidence interval for an odds ratio with .5 added to
each cell frequency. This measure of association assumes the group variable
is naturally dichotomous and the response variable is artificially
dichotomous.
}
\examples{
ci.biphi(.05, 46, 15, 100, 100)

# Should return:
#  Estimate         SE        LL       UL
# 0.4145733 0.07551281 0.2508866 0.546141


}
\references{
\insertRef{Ulrich2004}{statpsych}
}
