% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{ci.var.upper}
\alias{ci.var.upper}
\title{Upper confidence limit of variance}
\usage{
ci.var.upper(alpha, var, n)
}
\arguments{
\item{alpha}{alpha value for 1-alpha confidence (one-sided)}

\item{var}{estimated variance}

\item{n}{sample size}
}
\value{
Returns an upper limit (UL) variance planning value
}
\description{
Computes an upper limit for a population variance using an estimated variance
from a sample of size n in a prior study. The upper limit can be used as
a variance planning value in sample size functions that require a variance
planning value.
}
\examples{
ci.var.upper(.25, 15, 60)

# Should return:
#       UL
# 17.23264
 

}
