% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych2.R
\name{size.test.cor2}
\alias{size.test.cor2}
\title{Sample size for a test of equal Pearson or partial correlation in a 2-group
design}
\usage{
size.test.cor2(alpha, pow, cor1, cor2, s)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{cor1}{planning value of correlation}

\item{cor2}{planning value of correlation}

\item{s}{number of control variables}
}
\value{
Returns the required sample size per group
}
\description{
Computes the sample size required to test equality of two Pearson or partial
correlation with desired power in a 2-group design. Set s = 0 for a Pearson
correlation.
}
\examples{
size.test.cor2(.05, .8, .4, .2, 0)

# Should return:
#      Sample size per group
# [1,]                   325
 

}
