% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych3.R
\name{size.test.prop}
\alias{size.test.prop}
\title{Sample size for a test of a single proportion}
\usage{
size.test.prop(alpha, pow, p, h)
}
\arguments{
\item{alpha}{alpha level for hypothesis test}

\item{pow}{desired power}

\item{p}{planning value of proportion}

\item{h}{null hypothesis value of proportion}
}
\value{
Returns the required sample size
}
\description{
Computes the sample size required to test a population proportion with
desired power using a correction for continuity in a 1-group design.
}
\examples{
size.test.prop(.05, .9, .5, .3)

# Should return:
# Sample size
#          67


}
\references{
\insertRef{Fleiss2003}{statpsych}
}
