% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statpsych1.R
\name{spearmanbrown}
\alias{spearmanbrown}
\title{Computes the reliability of a scale with r2 measurements given the
reliability of a scale with r1 measurements}
\usage{
spearmanbrown(rel, r1, r2)
}
\arguments{
\item{rel}{reliability of the sum or average of r1 measurements}

\item{r1}{number of measurements in the orginal scale}

\item{r2}{number of measurements in the new scale}
}
\value{
Returns the reliability of the sum or average of r2 measurements
}
\description{
Computes the reliability of a scale that is the sum or average of r2
parallel measurements given the reliability of a scale that is the sum or
average of r1 parallel measurements. The "measurements" can be items,
trials, raters, or occasions.
}
\examples{
spearmanbrown(.6, 10, 20)

# Should return:
# Reliability of r2 measurements
#                            .75


}
